refactor(aicore/filtering): collapse public API into filters.py
Address review comment #14 ("can we have all public methods on a
filters.py? Better to read").

Merge _filters.py + _modules.py + _api.py into a single filters.py
(no underscore — signals the public surface). It now owns Severity,
ContentFilter, AzureContentFilter, LlamaGuard38bFilter, InputFiltering,
OutputFiltering, ContentFiltering, set_filtering, disable_filtering.
Tests import directly from .filters; the package __init__ stays a thin
re-export for the flat 'from sap_cloud_sdk.aicore import ...' path.

Internal-only files (_litellm_patch, exceptions) keep their separate
locations; they aren't part of the documented public surface.
